cleanup frontend, prep api for production
