+++ quadratic-api/package.json
@@ -1,6 +1,6 @@
 {
   "name": "quadratic-api",
-  "version": "1.0.0",
+  "version": "0.1.0",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -12,7 +12,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "David Kircos",
-  "license": "ISC",
+  "license": "MIT",
   "dependencies": {
     "@prisma/client": "^4.6.1",
     "@types/express": "^4.17.14",
@@ -21,6 +21,7 @@
     "express": "^4.18.2",
     "express-jwt": "^7.7.7",
     "express-rate-limit": "^6.7.0",
+    "helmet": "^6.0.1",
     "jwks-rsa": "^3.0.0",
     "nodemon": "^2.0.20",
     "openai": "^3.2.1",
