--- quadratic-api/prisma/schema.prisma
@@ -1,30 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
-}
-
-// Models
-
-model QUser {
-  id            Int     @id @default(autoincrement())
-  auth0_user_id String? @unique
-  QFile         QFile[]
-}
-
-model QFile {
-  id           Int      @id @default(autoincrement())
-  url          String   @unique @default(uuid())
-  user_owner   QUser    @relation(fields: [qUserId], references: [id])
-  name         String
-  data         Bytes?
-  created_date DateTime @default(now())
-  updated_date DateTime @default(now())
-  qUserId      Int
-}
