--- quadratic-api/src/types.ts
@@ -1,3 +1,0 @@
-import { QFile as QFileType } from '@prisma/client';
-
-export type QFile = QFileType;
